Revert "test(Increase-timeout-of-integrationt-test.): Increase timeout to 30 seconds to fix integration tests (#60)" (#89)
This reverts commit c4c7484670ffd9687321ce577d92722a31e635d0.

Revert the change to increase timeout of integ tests. The timeout was actually an issue because with the transport being used, the proxy creates session for every requests.

Because we are using stdio which is not aware of the sessionId, the client calls proxy with session ID, and proxy creates a new session.